Crate prompted

Prompting macros.

These are eventually intended to go into the standard library macros. At that point, parts of them will need to be reimplented for efficiency and compatibility.

Macros

eprompt

Same functionality as prompt!() except using stderr() instead of stdout().

input

If a [format!] describing a prompt is present, print it on stdout() and then flush. Then read a line from stdin() and return it after removing the line ending.

prompt

Same functionality as print!() except that stdout() is flushed at the end.

Functions

eflush

Flush standard error. Intended primarily for use by macros.

flush

Flush standard output. Intended primarily for use by macros.

read_line

Read a line from standard input. Removes a trailing newline if present. Intended primarily for use by macros.